LayoutTests/platform/mac/accessibility/aria-multiline-expected.txt

a
b
This tests that aria-multiline will change the role of a text control from a text field to a text area.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.role is 'AXRole: AXTextField'
PASS accessibilityController.focusedElement.role is 'AXRole: AXTextArea'
PASS successfullyParsed is true

TEST COMPLETE
